Extract a `Dependabot::Version` class
