+++ common/lib/dependabot.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
 module Dependabot
+  VERSION = "0.215.0"
 end
