+++ common/spec/dummy_package_manager/version.rb
@@ -1,9 +1,10 @@
 # frozen_string_literal: true
 
+require "dependabot/version"
 require "dependabot/utils"
 
 module DummyPackageManager
-  class Version < Gem::Version
+  class Version < Dependabot::Version
     def initialize(version)
       version = Version.remove_leading_v(version)
       super
