+++ git_submodules/lib/dependabot/git_submodules/version.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 
+require "dependabot/version"
 require "dependabot/utils"
 
 module Dependabot
   module GitSubmodules
-    class Version < Gem::Version
+    class Version < Dependabot::Version
     end
   end
 end
