+++ github_actions/lib/dependabot/github_actions/version.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 
+require "dependabot/version"
 require "dependabot/utils"
 
 module Dependabot
   module GithubActions
-    class Version < Gem::Version
+    class Version < Dependabot::Version
       def initialize(version)
         version = Version.remove_leading_v(version)
         super
