+++ nuget/lib/dependabot/nuget/version.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
+require "dependabot/version"
 require "dependabot/utils"
-require "rubygems_version_patch"
 
 # Dotnet pre-release versions use 1.0.1-rc1 syntax, which Gem::Version
 # converts into 1.0.1.pre.rc1. We override the `to_s` method to stop that
@@ -9,7 +9,7 @@ require "rubygems_version_patch"
 # Dotnet also supports build versions, separated with a "+".
 module Dependabot
   module Nuget
-    class Version < Gem::Version
+    class Version < Dependabot::Version
       VERSION_PATTERN = Gem::Version::VERSION_PATTERN + '(\+[0-9a-zA-Z\-.]+)?'
       ANCHORED_VERSION_PATTERN = /\A\s*(#{VERSION_PATTERN})?\s*\z/
