+++ terraform/lib/dependabot/terraform/version.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "dependabot/version"
+
 # Terraform pre-release versions use 1.0.1-rc1 syntax, which Gem::Version
 # converts into 1.0.1.pre.rc1. We override the `to_s` method to stop that
 # alteration.
@@ -8,7 +10,7 @@
 
 module Dependabot
   module Terraform
-    class Version < Gem::Version
+    class Version < Dependabot::Version
       def initialize(version)
         @version_string = version.to_s
         super
